questions list hook
