+++ src/pages/Room.tsx
@@ -5,6 +5,7 @@ import { Button } from '../components/Button'
 import { Question } from '../components/Question'
 import { RoomCode } from '../components/RoomCode'
 import { useAuth } from '../hooks/useAuth'
+import { UseRoom } from '../hooks/useRoom'
 import { database } from '../services/firebase'
 import '../styles/room.scss'
 
@@ -13,66 +14,17 @@ type RoomParams = {
     id: string;
 }
 
-type FirebaseQuestions = Record<string, {
-    author:{
-        name: string;
-        avatar: string;
-    }
-
-    content: string;
-    isAnswered: boolean;
-    isHighLighted: boolean
-}>
-
-type QuestionType = {
-    author: {
-        name: string;
-        avatar: string;
-    }
-    id: string;
-    content: string;
-    isAnswered: boolean;
-    isHighLighted: boolean
-}
 
 export function Room() {
+    const params = useParams<RoomParams>()
+    const roomId = params.id
 
+    const {title, questions} = UseRoom(roomId)
     
     const {user} = useAuth()
     
-    const params = useParams<RoomParams>()
-    const roomId = params.id
     
     const [newQuestion, setNewQuestion] = useState('')
-    const [questions, setQuestions] = useState<QuestionType[]>([])
-    const [title, setTitle] = useState('')
-    
-    console.log(questions)
-    useEffect(()=> {
-        const roomRef = database.ref(`rooms/${roomId}`)
-
-        roomRef.on('value', room => {
-            const databaseRoom = room.val()
-
-            const FirebaseQuestions: FirebaseQuestions = databaseRoom.questions ?? {}
-
-            const parsedQuestions = Object.entries(FirebaseQuestions).map(([key, value]) => {
-                return {
-                    id: key,
-                    author: value.author,
-                    content: value.content,
-                    isHighLighted: value.isHighLighted,
-                    isAnswered: value.isAnswered,
-                }
-            })
-            
-            setTitle(databaseRoom.title)
-            //@ts-ignore
-            setQuestions(parsedQuestions)
-        } )
-
-
-    }, [roomId])
 
     async function handleSendQuestion(event: FormEvent){
